refactor(rpc): merge Starknet read/write/trace traits into single StarknetApi
Collapse `StarknetApi`, `StarknetWriteApi`, and `StarknetTraceApi` into a single
`StarknetApi` trait sharing the existing `starknet` namespace. Server impls under
crates/rpc/rpc-server/src/starknet/{read,write,trace}.rs are merged into a single
api.rs (~480 lines) that preserves the inherent `add_*_tx` / `*_sync` helpers and
the `KatanaApiServer` impl that depends on them.

Both node assemblies (full + sequencer) drop two of their three `merge()` calls
into the JSON-RPC module. The katana-starknet client crate drops two of its three
imports (`StarknetWriteApiClient`, `StarknetTraceApiClient`).

BREAKING: Removes `StarknetWriteApi`, `StarknetTraceApi`, and their `*Server` /
`*Client` derivatives from the public surface of `katana-rpc-api`. Downstream
consumers should replace those imports with `StarknetApi`.
