+++ crates/node/full/src/lib.rs
@@ -29,7 +29,7 @@ use katana_pipeline::{Pipeline, PipelineHandle};
 use katana_pool::ordering::TipOrdering;
 use katana_provider::DbProviderFactory;
 use katana_rpc_api::katana::KatanaApiServer;
-use katana_rpc_api::starknet::{StarknetApiServer, StarknetTraceApiServer, StarknetWriteApiServer};
+use katana_rpc_api::starknet::StarknetApiServer;
 use katana_rpc_server::middleware::cors::Cors;
 use katana_rpc_server::starknet::{RpcCache, StarknetApi, StarknetApiConfig};
 use katana_rpc_server::{RpcServer, RpcServerHandle};
@@ -401,8 +401,6 @@ impl Node {
             }
 
             rpc_modules.merge(StarknetApiServer::into_rpc(starknet_api.clone()))?;
-            rpc_modules.merge(StarknetWriteApiServer::into_rpc(starknet_api.clone()))?;
-            rpc_modules.merge(StarknetTraceApiServer::into_rpc(starknet_api.clone()))?;
             rpc_modules.merge(KatanaApiServer::into_rpc(starknet_api.clone()))?;
         }
